fix!: refactor SQL scripts and imports
