+++ backend/src/prisma/cm/prisma.cm.module.ts
@@ -1,7 +1,6 @@
 import { Module } from "@nestjs/common";
 import { CaseManagementPrismaService } from "./prisma.cm.service";
-//Ignoring Sonar Warning on the line below since we control the prisma client.
-import { PrismaClient } from "../../../node_modules/.prisma/case_management"; // NOSONAR
+import { PrismaClient } from ".prisma/case_management";
 
 @Module({
   providers: [
