+++ docker-compose.yml
@@ -42,6 +42,7 @@ services:
       FLYWAY_PASSWORD: *POSTGRES_PASSWORD
       FLYWAY_PLACEHOLDERS_SHARED_PASSWORD: *SHARED_PASSWORD
       FLYWAY_PLACEHOLDERS_CM_PASSWORD: *CM_PASSWORD
+      FLYWAY_PLACEHOLDERS_INV_PASSWORD: *CM_PASSWORD
       FLYWAY_BASELINE_ON_MIGRATE: true
       FLYWAY_DEFAULT_SCHEMA: case_management
       FLYWAY_LOCATIONS: filesystem:./sql, filesystem:./sql-test
